Move Tron into separate file and
