+++ Sources/Bitcoin/Bitcoin.swift
@@ -67,23 +67,6 @@ public final class Litecoin: Bitcoin {
     }
 }
 
-public final class Tron: Bitcoin {
-    public override var coinType: Slip {
-        return .tron
-    }
-    public override var payToScriptHashAddressPrefix: UInt8 {
-        return 0x41
-    }
-
-    override open var coinPurpose: Purpose {
-        return .bip39
-    }
-
-    open override func address(for publicKey: PublicKey) -> Address {
-        return publicKey.tronAddress()
-    }
-}
-
 public final class Dash: Bitcoin {
     public override var coinType: Slip {
         return .dash
