+++ Tests/PurposeTests.swift
@@ -10,9 +10,12 @@ import TrustCore
 class PurposeTests: XCTestCase {
 
     func testPurpose() {
+        XCTAssertEqual(.bip39, Tron().coinPurpose)
+
         XCTAssertEqual(.bip49, Bitcoin().coinPurpose)
         XCTAssertEqual(.bip49, Litecoin().coinPurpose)
         XCTAssertEqual(.bip49, Dash().coinPurpose)
+
         XCTAssertEqual(.bip44, Ethereum().coinPurpose)
         XCTAssertEqual(.bip44, Wanchain().coinPurpose)
         XCTAssertEqual(.bip44, Vechain().coinPurpose)
